Fix P0 code review issues: schema alignment, Image.file, PUT→PATCH
- AI schemas: price fields nullable, bid_amount float, confidence pattern
- Mobile: Image.network→Image.file for local photos, PUT→PATCH endpoint
- Mobile: photo limit 20→10 (backend max), price validators for reserve/BIN
- Web: logout sends refresh_token body, 401 interceptor shared promise

+++ ai-service/app/services/fraud.py
@@ -19,7 +19,7 @@ VERY_HIGH_BID_THRESHOLD = 5_000_000  # 50,000 JOD — very high bid
 async def score_fraud(
     user_id: str,
     auction_id: str,
-    bid_amount: int,
+    bid_amount: float,
 ) -> FraudScoreResponse:
     """Compute a fraud risk score based on rule-based heuristics.
